float for percentage computation
Former-commit-id: 834f44a8c5bb3d83e792fd7b28259481e3cfa36b

+++ OpenBLAS/backend/sw/gemm_backend.h
@@ -531,11 +531,11 @@ static int gemm_backend_test (
         uint64_t time_prepare_action = timediff_usec(&etime_prepare_action,  &stime_prepare_action);
         uint64_t time_action_execute = timediff_usec(&etime_action_execution,  &stime_action_execution);
 	uint64_t time_total = time_card_allocation + time_attach_action + time_memory_allocation + time_prepare_action + time_action_execute;
-	VERBOSE3(stdout, "time card allocation (us): %lld, %lld\%\n",time_card_allocation, (time_card_allocation/time_total)*100);
-	VERBOSE3(stdout, "time attach action (us): %lld, %lld\%\n",time_attach_action, (time_attach_action/time_total)*100);
-	VERBOSE3(stdout, "time memory allocation (us): %lld, %lld\%\n",time_memory_allocation, (time_memory_allocation/time_total)*100);
-	VERBOSE3(stdout, "time action prepare(us): %lld, %lld\%\n",time_prepare_action, (time_prepare_action/time_total)*100);
-	VERBOSE3(stdout, "time action execute(us): %lld, %lld\%\n",time_action_execute, (time_action_execute/time_total)*100);
+	VERBOSE3(stdout, "time card allocation (us): %lld, %lld\%\n",time_card_allocation, (float)(time_card_allocation/time_total)*100);
+	VERBOSE3(stdout, "time attach action (us): %lld, %lld\%\n",time_attach_action, (float)(time_attach_action/time_total)*100);
+	VERBOSE3(stdout, "time memory allocation (us): %lld, %lld\%\n",time_memory_allocation, (float)(time_memory_allocation/time_total)*100);
+	VERBOSE3(stdout, "time action prepare(us): %lld, %lld\%\n",time_prepare_action, (float)(time_prepare_action/time_total)*100);
+	VERBOSE3(stdout, "time action execute(us): %lld, %lld\%\n",time_action_execute, (float)(time_action_execute/time_total)*100);
 
     }
